Revise memory diagnostics to use task handle not task name
In addition, aggregation of the sizes of blocks allocated by each task
to get a total is now done in new mem_debug_malloc_dump_totals() so it
is not necessary to dump info for each individual heap block into an
array for the app code.  Only the blocks belonging to specific tasks
will be dumped when those tasks are selected by the user for a
detailed list of blocks.

This change in the memory diagnostic API is indicated by a change in
the conditional configENABLE_MEMORY_DEBUG_DUMP from 1 to 2.