Update cache when changing log levels so change is effective
If any log messages with a particular tag string had already been
output before the call was made to adjust the logging level then the
new setting would not take effect because the level was cached.

This change updates the cached level to the new value and also finds
and updates the existing entry in the full uncached list rather than
adding a new entry to the end of the list.